It works!!!
for tiny files

+++ app/src/main/kotlin/xin/z7workbench/recipie/entity/Entity.kt
@@ -9,6 +9,6 @@ data class ServerMessage(val Object: String, val ToUser: String, val FromUser: S
 
 data class FileInfoMessage(val Object: String, val ToUser: String, val FromUser: String,
                            val CreateTime: String, val MsgType: String, val FileName: String,
-                           val FileSize: Long, val MsgID: Int)
+                           val FileSize: Int, val MsgID: Int)
 
 data class FileMessage(val MsgType: String, val MsgID: Int, val Content: String)
